reverted app, ran build, added return types
